requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.1
python-docx==1.0.1
PyPDF2==3.0.1
beautifulsoup4==4.12.2
markdown==3.5.1
python-jose==3.3.0
pydantic==2.5.2
numpy==1.26.2
pandas==2.1.3
tqdm==4.66.1
pytest==7.4.3
httpx==0.25.2
streamlit==1.31.1 